+++ src/common/helpers/StringHelper.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @link https://github.com/black-lamp/yii2-slider
+ * @copyright Copyright (c) Vladimir Kuprienko
+ * @license BSD 3-Clause License
+ */
+
+namespace bl\slider\common\helpers;
+
+use bl\slider\common\helpers\base\BaseStringHelper;
+
+/**
+ * StringHelper provides a set of static methods for work with strings
+ *
+ * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ */
+class StringHelper extends BaseStringHelper
+{
+
+}